Implemented all Helper functions + treelib.ancestors patch
By Helper functions I mean what we were calling "wrapper" functions. I
find this name is more true to the purpose of these functions.

New in GUIVersion1.py:

class Helper: This is just a class for the sake of differentiating
functions who's only purpose is to make the code easier to read.

Helper : Functions:
getCommonRoot: Takes as arguments (tree,node,node). Returns the id of
the closest common root between two nodes in a given tree.
getCommonSubTree: Takes as arguments (tree,node,node). Returns a subtree
with root the closest common root between two nodes in a given tree.

"treelib.ancestors" patch: I noticed that the ancestors did not include
the node for which we are asking what its ancestors are. While this
seemed to originally make sense, it caused some problems when
implementing getCommonSubTree because inputting a node and its parent
did not return a tree with that parent as its root, even though this is
the smallest possible sub-tree containing both nodes.